requirements.txt
requests
tqdm
pymupdf
aiohttp
supabase
python-dotenv
fastapi
uvicorn
sse-starlette
